2nd commit

+++ src/main/java/com/wealth/stock/controller/optionGrowController.java
@@ -57,7 +57,7 @@ public class optionGrowController {
                         , istTime.getHour(), istTime.getMinute(), istTime.getSecond());
                 as.optionChain.optionChains.forEach(optionChain -> {
                     optionChain.timeStamp = date;
-                    optionChain.pcr=optionChain.putOption.openInterest/optionChain.callOption.openInterest;
+                    optionChain.pcr=(double) optionChain.putOption.openInterest/optionChain.callOption.openInterest;
                     optionChainRepository.save(optionChain);
                 });
                 Future future = new Future();
